--- src/main/java/roomescape/controller/CreateReservationRequest.java
@@ -1,7 +1,0 @@
-package roomescape.controller;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-record CreateReservationRequest(LocalDate date, String name, LocalTime time) {
-}
