--- src/main/java/roomescape/controller/CreateReservationTimeRequest.java
@@ -1,6 +1,0 @@
-package roomescape.controller;
-
-import java.time.LocalTime;
-
-record CreateReservationTimeRequest(LocalTime startAt) {
-}
