+++ src/main/java/roomescape/controller/ReservationController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.request.CreateReservationRequest;
+import roomescape.controller.response.ReservationResponse;
 
 @RestController
 @RequestMapping("/reservations")
