+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.request.CreateReservationTimeRequest;
+import roomescape.controller.response.ReservationTimeResponse;
 
 @RestController
 @RequestMapping("/times")
