+++ src/main/java/roomescape/controller/request/CreateReservationRequest.java
@@ -1,0 +1,7 @@
+package roomescape.controller.request;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record CreateReservationRequest(LocalDate date, String name, LocalTime time) {
+}
