+++ src/main/java/roomescape/controller/request/CreateReservationTimeRequest.java
@@ -1,0 +1,6 @@
+package roomescape.controller.request;
+
+import java.time.LocalTime;
+
+public record CreateReservationTimeRequest(LocalTime startAt) {
+}
